reapply techpack: display: Use backlight brightness for HBM after a certain threshold
Screen blinks for a milisecond when activating/deactivating HBM in the brightness range of 256-486.
Since the screen is bright enough for fingerprint scans after 256, adjust the HBM mask brightness
to the screen brightness to prevent screen blinking during high brightness FOD scans.

Samsung also has a dimming table up to 255 so it should be fine.

Change-Id: I1ff58d08e7f53d0d9c4a6f91f5d4e9afc5333cc4